moved submission by countries CSV to superuser_stats
